+++ vcl/aqua/source/app/salinst.cxx
@@ -1303,7 +1303,7 @@ NSImage* CreateNSImage( const Image& rImage )
         NSGraphicsContext* pContext = [NSGraphicsContext currentContext];
         CGContextRef rCGContext = reinterpret_cast<CGContextRef>([pContext graphicsPort]);
 
-        const CGRect aDstRect = { {0, 0}, { aSize.Width(), aSize.Height() } };
+        const CGRect aDstRect = { {0, 0}, { static_cast<CGFloat>(aSize.Width()), static_cast<CGFloat>(aSize.Height()) } };
         CGContextDrawImage( rCGContext, aDstRect, xImage );
 
         [pImage unlockFocus];
